requirements.txt
iopath==0.1.10
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.3
PyYAML==6.0
torch==1.12.0
